config/linux/ipu6: ov01a1s configs: Adjust for new 1288x800 ov01a1s kernel driver output size
There is a patch series pending upstream:

https://lore.kernel.org/linux-media/20251014174033.20534-1-hansg@kernel.org/

To add ov01a1s support to the mainline kernel. The mainline support will
have a slightly different output-size with a maximum output width of 1288
as the mainline code keeps a 4x4 border reserved to be able to shift
the window around to keep the Bayer pattern unchanged when doing flipping.

And with the mainline driver the output height must be a multiple of 4
because the RGBI Bayer patern on the ov01a1s sensor repeats every 4 lines.

There also is a pending pull-request to change the output-size of
the ov01a1s driver in intel-ipu6-drivers to match the mainline driver:

https://github.com/intel/ipu6-drivers/pull/395

This commit adjusts the graph_settings_ov01a1s.xml and ov01a1s-uf.xml
config files to work with the new 1288x800 output size.

Signed-off-by: Hans de Goede <hansg@kernel.org>